pyproject.toml
[tool.poetry]
name = "venv"
version = "0.1.0"
description = ""
authors = ["dedoxd2 <dedo.xd33@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
pytest-django = "^4.5.2"
django-pytest = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
